GestationPeriodinYears datatype modification

+++ NPD/EmcureNPD.Data.DataAccess/Entity/PidfFinance.cs
@@ -30,7 +30,7 @@ namespace EmcureNPD.Data.DataAccess.Entity
         public string ApprovalPeriodinDays { get; set; }
         public DateTime? ApprovalDate { get; set; }
         public DateTime? ProductLaunchDate { get; set; }
-        public int? GestationPeriodinYears { get; set; }
+        public decimal? GestationPeriodinYears { get; set; }
         public decimal? MarketShareErosionrate { get; set; }
         public decimal? PriceErosion { get; set; }
         public string EscalationinCogs { get; set; }
